Don't block Googlebot from crawling JavaScript and CSS
The robots exclusion protocol is not a complete security solution - it is not a substitute for password protection. Thus the only real gain by blocking all of the above is a slightly narrower hole where holes exist (they shouldn't exist in the first place), and from a non-security perspective there can be minor freeing of server resources by having less bots crawling the site. HOWEVER, these robots entries are detrimental to Search Engine Optimisation (SEO) and that should be the over-riding factor considering the pros and cons involved.

Notice that WordPress doesn't have this problem because it puts plugins, templates and image uploads etc in the /wp-content/ folder which is not blocked by robots.txt by default.

For SEO, all client-side dependables should be free for robots to crawl, to assess site quality, to determine if images and styles are good and relevant, and to ensure cloaking (hidden content) and sneaky redirects are not present. This is critical to SEO. For evidence I'm referring to Matt Cutts on the official YouTube channel of Google Webmaster Help:

Don't block Googlebot from crawling JavaScript and CSS
http://www.youtube.com/watch?v=B9BWbruCiDc